+++ package.json
@@ -51,6 +51,7 @@
     "postcss": "^8.5.8",
     "tailwindcss": "^3.4.17",
     "typescript": "~5.9.3",
+    "unplugin-vue-components": "^30.0.0",
     "vite": "^8.0.1",
     "vue-tsc": "^3.2.5"
   }
